forge fmt

+++ tests/Contracts/src/Lib.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.13;
 
 library Lib {
-    function doSomething(uint a, uint b) external pure returns (uint) {
+    function doSomething(uint256 a, uint256 b) external pure returns (uint256) {
         return a + b;
     }
 }
